fix: mixup between buffer item count and byte sizes and offsets
Additionally, this makes the type-safe version consistent with
conventional Rust containers such as slices and `Vec`s and reduces the
need for lossy integer conversions as conversions now mostly occur to
larger target integer types.

This is prominent for wasm where `usize: u32` and casting back and forth
between `u64` and `u32` is lossy.

Furthermore, this makes commit makes bounds checking stricter to ensure
that no overread and overwrite, as well as uninitialized reads or
incorrectly bound ranges occur,

+++ crates/renderer/src/collect.rs
@@ -258,7 +258,7 @@ impl RendererCollect {
         }
 
         tracing::debug!("Resizing collect command buffer to {primitives_count}");
-        output.commands.resize(gpu, primitives_count as u64, true);
+        output.commands.resize(gpu, primitives_count as usize, true);
 
         assert_eq!(
             input_primitives.total_len(),
@@ -366,7 +366,7 @@ impl CollectCountStagingBuffers {
     }
 
     #[cfg(any(target_os = "macos", target_os = "unknown"))]
-    fn take_buffer(&self, gpu: &Gpu, size: u64) -> TypedBuffer<u32> {
+    fn take_buffer(&self, gpu: &Gpu, size: usize) -> TypedBuffer<u32> {
         match self.buffers.lock().pop() {
             Some(mut buffer) => {
                 buffer.resize(gpu, size, false);
